refactor(api): reduce root router to composition and split bounded routers
Extract health, rag, docs and index endpoints into dedicated router modules and move runtime wiring/composition helpers to app/wiring.py.

Migrate ingest schemas into app/schemas and adjust unit tests to patch the new seams (wiring and bounded routers) while preserving endpoint behavior.

Update architecture docs to reflect api_router composition-only role and new router/wiring structure.
