+++ tests/unit/app/test_api_router_endpoints_extra.py
@@ -1,7 +1,7 @@
 # tests/unit/app/test_api_router_endpoints_extra.py
 from types import SimpleNamespace
 
-from local_rag_backend.app import api_router as api
+from local_rag_backend.app.routers import health as health_router
 from local_rag_backend.settings import settings
 
 
@@ -16,7 +16,7 @@ async def test_health_db_failure(asgi_client, monkeypatch):
         def __exit__(self, exc_type, exc, tb):
             return False
 
-    monkeypatch.setattr(api.db_base, "engine", SimpleNamespace(connect=lambda: DummyConn()))
+    monkeypatch.setattr(health_router.db_base, "engine", SimpleNamespace(connect=lambda: DummyConn()))
     r = await asgi_client.get("/api/health")
     assert r.status_code == 503
     assert "Database connection failed" in r.json()["detail"]
@@ -31,14 +31,14 @@ async def test_ready_not_ready_db_and_no_llm(asgi_client, monkeypatch):
         def __exit__(self, *a):
             return False
 
-    monkeypatch.setattr(api.db_base, "engine", SimpleNamespace(connect=lambda: DummyConn()))
+    monkeypatch.setattr(health_router.db_base, "engine", SimpleNamespace(connect=lambda: DummyConn()))
     monkeypatch.setattr(settings, "openai_api_key", None, raising=False)
     monkeypatch.setattr(settings, "ollama_enabled", False, raising=False)
 
     async def _override():
         return None
 
-    monkeypatch.setattr(api, "get_rag_service", _override, raising=True)
+    monkeypatch.setattr(health_router, "get_rag_service", _override, raising=True)
     r = await asgi_client.get("/api/ready")
     assert r.status_code == 503
     detail = r.json()["detail"]
@@ -57,7 +57,7 @@ async def test_ollama_health_ok(asgi_client, monkeypatch):
         def raise_for_status(self):
             return None
 
-    monkeypatch.setattr(api.requests, "get", lambda *a, **k: Resp())
+    monkeypatch.setattr(health_router.requests, "get", lambda *a, **k: Resp())
     r = await asgi_client.get("/api/health/ollama")
     assert r.status_code == 200
     assert r.json()["status"] == "ok"
@@ -68,8 +68,8 @@ async def test_ollama_health_fail(asgi_client, monkeypatch):
         pass
 
     def boom(*a, **k):
-        raise api.requests.exceptions.RequestException("oops")
+        raise health_router.requests.exceptions.RequestException("oops")
 
-    monkeypatch.setattr(api.requests, "get", boom)
+    monkeypatch.setattr(health_router.requests, "get", boom)
     r = await asgi_client.get("/api/health/ollama")
     assert r.status_code == 503
