+++ tests/unit/app/test_docs_index_ports_wiring.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from local_rag_backend.app import api_router as api
+from local_rag_backend.app import wiring
 
 if TYPE_CHECKING:
     import pytest
@@ -32,15 +32,15 @@ def test_docs_mutation_ports_bind_current_router_symbols(monkeypatch: pytest.Mon
     def _dummy_delete_external(**kwargs: object) -> tuple[int, int | None, list[str], int, bool]:
         return 0, None, [], 0, False
 
-    monkeypatch.setattr(api, "SqlDocumentStorage", DummyRepo, raising=True)
-    monkeypatch.setattr(api, "FaissVectorStorage", _dummy_vec_factory, raising=True)
-    monkeypatch.setattr(api, "precompute_vectors_for_changed_items", _dummy_precompute, raising=True)
-    monkeypatch.setattr(api, "sync_dense_after_upsert", _dummy_sync, raising=True)
-    monkeypatch.setattr(api, "rebuild_index_from_db", _dummy_rebuild, raising=True)
-    monkeypatch.setattr(api, "delete_documents_multi_store", _dummy_delete_docs, raising=True)
-    monkeypatch.setattr(api, "delete_external_ids_multi_store", _dummy_delete_external, raising=True)
+    monkeypatch.setattr(wiring, "SqlDocumentStorage", DummyRepo, raising=True)
+    monkeypatch.setattr(wiring, "FaissVectorStorage", _dummy_vec_factory, raising=True)
+    monkeypatch.setattr(wiring, "precompute_vectors_for_changed_items", _dummy_precompute, raising=True)
+    monkeypatch.setattr(wiring, "sync_dense_after_upsert", _dummy_sync, raising=True)
+    monkeypatch.setattr(wiring, "rebuild_index_from_db", _dummy_rebuild, raising=True)
+    monkeypatch.setattr(wiring, "delete_documents_multi_store", _dummy_delete_docs, raising=True)
+    monkeypatch.setattr(wiring, "delete_external_ids_multi_store", _dummy_delete_external, raising=True)
 
-    ports = api._docs_mutation_ports()
+    ports = wiring.docs_mutation_ports()
 
     assert isinstance(ports.doc_repo_factory(), DummyRepo)
     assert ports.build_upsert_doc is DummyRepo.UpsertDoc
@@ -65,12 +65,12 @@ def test_index_mutation_ports_bind_current_router_symbols(monkeypatch: pytest.Mo
     def _dummy_rebuild(**kwargs: object) -> int:
         return 0
 
-    monkeypatch.setattr(api, "SqlDocumentStorage", DummyRepo, raising=True)
-    monkeypatch.setattr(api, "FaissVectorStorage", _dummy_vec_factory, raising=True)
-    monkeypatch.setattr(api, "purge_index_artifacts", _dummy_purge, raising=True)
-    monkeypatch.setattr(api, "rebuild_index_from_db", _dummy_rebuild, raising=True)
+    monkeypatch.setattr(wiring, "SqlDocumentStorage", DummyRepo, raising=True)
+    monkeypatch.setattr(wiring, "FaissVectorStorage", _dummy_vec_factory, raising=True)
+    monkeypatch.setattr(wiring, "purge_index_artifacts", _dummy_purge, raising=True)
+    monkeypatch.setattr(wiring, "rebuild_index_from_db", _dummy_rebuild, raising=True)
 
-    ports = api._index_mutation_ports()
+    ports = wiring.index_mutation_ports()
 
     assert isinstance(ports.doc_repo_factory(), DummyRepo)
     assert ports.vector_repo_factory is _dummy_vec_factory
