+++ tests/unit/app/test_docs_upsert_endpoint.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from local_rag_backend.app import api_router as api
+from local_rag_backend.app import wiring
 from local_rag_backend.infrastructure.persistence.sqlalchemy.sql_ import SqlDocumentStorage
 from local_rag_backend.settings import settings
 
@@ -90,8 +90,8 @@ async def test_upsert_docs_dense_updates_only_changed_content(
     dummy_embedder = DummyEmbedder()
     dummy_vec = DummyVec()
 
-    monkeypatch.setattr(api, "OpenAIEmbedder", lambda *a, **k: dummy_embedder)
-    monkeypatch.setattr(api, "FaissVectorStorage", lambda *a, **k: dummy_vec)
+    monkeypatch.setattr(wiring, "OpenAIEmbedder", lambda *a, **k: dummy_embedder)
+    monkeypatch.setattr(wiring, "FaissVectorStorage", lambda *a, **k: dummy_vec)
 
     payload = {"docs": [{"external_id": "doc-1", "content": "hello"}]}
     r1 = await asgi_client.post("/api/docs/upsert", json=payload)
